refactor: MenuResponse dto 추가

+++ src/main/java/kitchenpos/dto/request/MenuProductRequest.java
@@ -2,13 +2,12 @@ package kitchenpos.dto.request;
 
 public class MenuProductRequest {
     private Long productId;
-    private int quantity;
-
+    private long quantity;
 
     public MenuProductRequest() {
     }
 
-    public MenuProductRequest(final Long productId, final int quantity) {
+    public MenuProductRequest(final Long productId, final long quantity) {
         this.productId = productId;
         this.quantity = quantity;
     }
@@ -17,7 +16,7 @@ public class MenuProductRequest {
         return productId;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 }
